add loading status to carousel
